Architectural Design of this Privacy Policies and Permissions is given below
 Research and Documentation of Privacy Policies and Permissions for Common Applications in Honda Android Automotive OS
The privacy policies and permissions of frequently used Honda Android Automotive OS applications are examined in this document, with a particular emphasis on location data. We are going to look at the kinds of data that are gathered, how they are used, and how the privacy policies of certain applications are handled. We will also evaluate the breadth and depth of possible privacy threats to users, contrasting our results with industry best practices and regulatory requires like GDPR.